Delete register agent files from controller
