--- plugins/main/public/controllers/register-agent/components/group-input/group-input.scss
@@ -1,5 +1,0 @@
-.registerAgentLabels {
-  font-weight: 700;
-  font-size: 12px;
-  line-height: 20px;
-}
