--- plugins/main/public/controllers/register-agent/containers/register-agent/register-agent.scss
@@ -1,27 +1,0 @@
-.register-agent-wizard-container {
-  box-sizing: border-box;
-  min-height: 1271px;
-  margin-top: 44px;
-  background: #ffffff;
-  border: 1px solid rgba(52, 55, 65, 0.2);
-  max-width: 1030px;
-  padding-left: 72px;
-  padding-right: 63px;
-}
-
-.register-agent-wizard-title {
-  margin-top: 51px;
-  margin-bottom: 51px;
-  font-style: normal;
-  font-weight: 400;
-  font-size: 30px;
-  line-height: 36px;
-  display: flex;
-  justify-content: center;
-}
-
-.register-agent-wizard-close {
-  display: flex;
-  margin-top: 17px;
-  float: right;
-}
