--- plugins/main/public/controllers/register-agent/containers/register-agent/register-agent.tsx
@@ -1,248 +1,0 @@
-import React, { useState, useEffect } from 'react';
-import {
-  EuiFlexGroup,
-  EuiFlexItem,
-  EuiPanel,
-  EuiTitle,
-  EuiButtonEmpty,
-  EuiPage,
-  EuiPageBody,
-  EuiSpacer,
-  EuiProgress,
-  EuiButton,
-} from '@elastic/eui';
-import { WzRequest } from '../../../../react-services/wz-request';
-import { UI_LOGGER_LEVELS } from '../../../../../common/constants';
-import { UI_ERROR_SEVERITIES } from '../../../../react-services/error-orchestrator/types';
-import { ErrorHandler } from '../../../../react-services/error-management';
-import { getMasterRemoteConfiguration } from '../../../agent/components/register-agent-service';
-import './register-agent.scss';
-import { Steps } from '../steps/steps';
-import { InputForm } from '../../../../components/common/form';
-import { getGroups } from '../../services/register-agent-services';
-import { useForm } from '../../../../components/common/form/hooks';
-import { FormConfiguration } from '../../../../components/common/form/types';
-import { useSelector } from 'react-redux';
-import {
-  withReduxProvider,
-  withGlobalBreadcrumb,
-  withUserAuthorizationPrompt,
-} from '../../../../components/common/hocs';
-import GroupInput from '../../components/group-input/group-input';
-import { OsCard } from '../../components/os-selector/os-card/os-card';
-import {
-  validateServerAddress,
-  validateAgentName,
-} from '../../utils/validations';
-import { compose } from 'redux';
-import { endpointSumary } from '../../../../utils/applications';
-import { getCore } from '../../../../kibana-services';
-
-interface IRegisterAgentProps {
-  getWazuhVersion: () => Promise<string>;
-  addNewAgent: (agent: any) => Promise<any>;
-  reload: () => void;
-}
-
-export const RegisterAgent = compose(
-  withReduxProvider,
-  withGlobalBreadcrumb([
-    { text: endpointSumary.title, href: '#/agents-preview' },
-    { text: 'Deploy new agent' },
-  ]),
-  withUserAuthorizationPrompt([
-    [{ action: 'agent:read', resource: 'agent:group:*' }],
-  ]),
-)(({ getWazuhVersion, addNewAgent, reload }: IRegisterAgentProps) => {
-  const configuration = useSelector(
-    (state: { appConfig: { data: any } }) => state.appConfig.data,
-  );
-  const [wazuhVersion, setWazuhVersion] = useState('');
-  const [haveUdpProtocol, setHaveUdpProtocol] = useState<boolean | null>(false);
-  const [loading, setLoading] = useState(false);
-  const [wazuhPassword, setWazuhPassword] = useState('');
-  const [groups, setGroups] = useState([]);
-  const [needsPassword, setNeedsPassword] = useState<boolean>(false);
-
-  const initialFields: FormConfiguration = {
-    operatingSystemSelection: {
-      type: 'custom',
-      initialValue: '',
-      component: props => {
-        return <OsCard {...props} />;
-      },
-      options: {
-        groups,
-      },
-    },
-    serverAddress: {
-      type: 'text',
-      initialValue: configuration['enrollment.dns'] || '',
-      validate: validateServerAddress,
-    },
-    agentName: {
-      type: 'text',
-      initialValue: '',
-      validate: validateAgentName,
-    },
-
-    agentGroups: {
-      type: 'custom',
-      initialValue: [],
-      component: props => {
-        return <GroupInput {...props} />;
-      },
-      options: {
-        groups,
-      },
-    },
-  };
-
-  const form = useForm(initialFields);
-
-  const getRemoteConfig = async () => {
-    const remoteConfig = await getMasterRemoteConfiguration();
-    if (remoteConfig) {
-      setHaveUdpProtocol(remoteConfig.isUdp);
-    }
-  };
-
-  const getAuthInfo = async () => {
-    try {
-      const result = await WzRequest.apiReq(
-        'GET',
-        '/agents/000/config/auth/auth',
-        {},
-      );
-      return (result.data || {}).data || {};
-    } catch (error) {
-      ErrorHandler.handleError(error);
-    }
-  };
-
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const wazuhVersion = await getWazuhVersion();
-        await getRemoteConfig();
-        const authInfo = await getAuthInfo();
-        // get wazuh password configuration
-        let wazuhPassword = '';
-        const needsPassword = (authInfo.auth || {}).use_password === 'yes';
-        if (needsPassword) {
-          wazuhPassword =
-            configuration['enrollment.password'] ||
-            authInfo['authd.pass'] ||
-            '';
-        }
-        const groups = await getGroups();
-        setNeedsPassword(needsPassword);
-        setWazuhPassword(wazuhPassword);
-        setWazuhVersion(wazuhVersion);
-        setGroups(groups);
-        setLoading(false);
-      } catch (error) {
-        setWazuhVersion(wazuhVersion);
-        setLoading(false);
-        const options = {
-          context: 'RegisterAgent',
-          level: UI_LOGGER_LEVELS.ERROR,
-          severity: UI_ERROR_SEVERITIES.BUSINESS,
-          display: true,
-          store: false,
-          error: {
-            error: error,
-            message: error.message || error,
-            title: error.name || error,
-          },
-        };
-        ErrorHandler.handleError(error, options);
-      }
-    };
-
-    fetchData();
-  }, []);
-
-  const osCard = (
-    <InputForm {...form.fields.operatingSystemSelection}></InputForm>
-  );
-
-  return (
-    <div>
-      <EuiPage restrictWidth='1000px' style={{ background: 'transparent' }}>
-        <EuiPageBody>
-          <EuiFlexGroup>
-            <EuiFlexItem>
-              <EuiPanel className='register-agent-wizard-container'>
-                <div className='register-agent-wizard-close'>
-                  <EuiButtonEmpty
-                    size='s'
-                    href={getCore().application.getUrlForApp(
-                      'endpoints-summary',
-                      {
-                        path: '#/agents-preview',
-                      },
-                    )}
-                    iconType='cross'
-                  >
-                    Close
-                  </EuiButtonEmpty>
-                </div>
-                <EuiFlexGroup>
-                  <EuiFlexItem>
-                    <EuiTitle>
-                      <h2 className='register-agent-wizard-title'>
-                        Deploy new agent
-                      </h2>
-                    </EuiTitle>
-                  </EuiFlexItem>
-                </EuiFlexGroup>
-                <EuiSpacer />
-                {loading ? (
-                  <>
-                    <EuiFlexItem>
-                      <EuiProgress size='xs' color='primary' />
-                    </EuiFlexItem>
-                    <EuiSpacer></EuiSpacer>
-                  </>
-                ) : (
-                  <EuiFlexItem>
-                    <Steps
-                      form={form}
-                      needsPassword={needsPassword}
-                      wazuhPassword={wazuhPassword}
-                      osCard={osCard}
-                      connection={{
-                        isUDP: haveUdpProtocol ? true : false,
-                      }}
-                    />
-                  </EuiFlexItem>
-                )}
-                <EuiFlexGroup
-                  justifyContent='flexEnd'
-                  style={{ marginRight: '0.3rem' }}
-                >
-                  <EuiFlexItem grow={false}>
-                    <EuiButton
-                      className='close-button'
-                      fill
-                      color='primary'
-                      href={getCore().application.getUrlForApp(
-                        'endpoints-summary',
-                        {
-                          path: '#/agents-preview',
-                        },
-                      )}
-                    >
-                      Close
-                    </EuiButton>
-                  </EuiFlexItem>
-                </EuiFlexGroup>
-              </EuiPanel>
-            </EuiFlexItem>
-          </EuiFlexGroup>
-        </EuiPageBody>
-      </EuiPage>
-    </div>
-  );
-});
