--- plugins/main/public/controllers/register-agent/index.tsx
@@ -1,1 +1,0 @@
-export { RegisterAgent } from './containers/register-agent/register-agent';
